+++ src/main/kotlin/io/github/facilityapi/intellij/tools/FsdChooseByNameContributor.kt
@@ -15,7 +15,7 @@ class FsdChooseByNameContributor : ChooseByNameContributor {
         name: String,
         pattern: String,
         project: Project,
-        includeNonProjectItems: Boolean
+        includeNonProjectItems: Boolean,
     ): Array<NavigationItem> {
         return findTypeDefinitions(project, name)
             .filterIsInstance<NavigationItem>()